move ChartsThemeProvider, revert generateChartsTheme debug
